api/requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
cffi==1.14.3
Django==3.1.1
djangorestframework==3.11.1
kombu==4.6.11
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
rabbitmq==0.2.0
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
gunicorn==20.0.4
